backend/requirements.txt
fastapi==0.115.1
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
redis==3.5.3
pyjwt==2.1.0
passlib==1.7.4
python-multipart==0.0.5
boto3==1.18.49
minio==7.1.0
wechatpy==1.8.7
qrcode==7.3.1
Pillow==8.3.2
python-jose==3.3.0
user-agents==2.2.0
alipay-sdk-python==3.1.6
fastapi-mail==1.4.1
pycryptodome==3.20.0
pydantic==2.10.4